requirements.txt
lightgbm==3.3.5
matplotlib==3.5.2
nltk==3.7
numpy==1.24.3
numpy==1.23.0
pandas==1.4.4
Pillow==9.2.0
Pillow==10.2.0
scikit_learn==1.0.2
textblob==0.17.1
wordcloud==1.9.2
